requirements.txt
black
click
Flask==2.2.2
requires.io==0.2.6
Pillow==9.2.0
pickleshare==0.7.5
tensorflow==2.12.0
scikit-learn==1.3.0
requests==2.31.0
cloudinary==1.33.0
gunicorn==20.1.0
numpy==1.22.2
